Fixed several bugs on AVX/SSE mathematical functions

+++ 12.1.3/tests/test_math.c
@@ -1665,10 +1665,10 @@ check_uni_f32 (float (*f) (const float), float (*g) (const float),
 {
   float x, xm, dx, f0, g0;
   int i, n1;
-  n1 = n - 1;
+  n1 = (int) n - 1;
   xm = 0.5f * (xmin + xmax);
   dx = (xmax - xm) / (float) n1;
-  for (i = 0; i < n; ++i)
+  for (i = 0; i < (int) n; ++i)
     {
       if (i == 0)
         x = xmin;
@@ -1781,10 +1781,10 @@ check_uni_f64 (double (*f) (const double), double (*g) (const double),
 {
   double x, xm, dx, f0, g0;
   int i, n1;
-  n1 = n - 1;
+  n1 = (int) n - 1;
   xm = 0.5 * (xmin + xmax);
   dx = (xmax - xm) / (double) n1;
-  for (i = 0; i < n; ++i)
+  for (i = 0; i < (int) n; ++i)
     {
       if (i == 0)
         x = xmin;
@@ -1887,10 +1887,10 @@ check_uni_4xf32 (__m128 (*f) (const __m128), float (*g) (const float),
   float x[4] JB_ALIGNED, f0[4] JB_ALIGNED, g0[4] JB_ALIGNED;
   float xm, dx;
   int i, j, n1;
-  n1 = n - 1;
+  n1 = (int) n - 1;
   xm = 0.5f * (xmin + xmax);
   dx = (xmax - xm) / (float) n1;
-  for (i = 0; i < n;)
+  for (i = 0; i < (int) n;)
     {
       for (j = 0; j < 4; ++j, ++i)
         {
@@ -2002,10 +2002,10 @@ check_uni_2xf64 (__m128d (*f) (const __m128d), double (*g) (const double),
   double x[2] JB_ALIGNED, f0[2] JB_ALIGNED, g0[2] JB_ALIGNED;
   double xm, dx;
   int i, j, n1;
-  n1 = n - 1;
+  n1 = (int) n - 1;
   xm = 0.5 * (xmin + xmax);
   dx = (xmax - xm) / (double) n1;
-  for (i = 0; i < n;)
+  for (i = 0; i < (int) n;)
     {
       for (j = 0; j < 2; ++j, ++i)
         {
@@ -2121,10 +2121,10 @@ check_uni_8xf32 (__m256 (*f) (const __m256), float (*g) (const float),
   float x[8] JB_ALIGNED, f0[8] JB_ALIGNED, g0[8] JB_ALIGNED;
   float xm, dx;
   int i, j, n1;
-  n1 = n - 1;
+  n1 = (int) n - 1;
   xm = 0.5f * (xmin + xmax);
   dx = (xmax - xm) / (float) n1;
-  for (i = 0; i < n;)
+  for (i = 0; i < (int) n;)
     {
       for (j = 0; j < 8; ++j, ++i)
         {
@@ -2236,10 +2236,10 @@ check_uni_4xf64 (__m256d (*f) (const __m256d), double (*g) (const double),
   double x[4] JB_ALIGNED, f0[4] JB_ALIGNED, g0[4] JB_ALIGNED;
   double xm, dx;
   int i, j, n1;
-  n1 = n - 1;
+  n1 = (int) n - 1;
   xm = 0.5 * (xmin + xmax);
   dx = (xmax - xm) / (double) n1;
-  for (i = 0; i < n;)
+  for (i = 0; i < (int) n;)
     {
       for (j = 0; j < 4; ++j, ++i)
         {
@@ -2355,10 +2355,10 @@ check_uni_16xf32 (__m512 (*f) (const __m512), float (*g) (const float),
   float x[16] JB_ALIGNED, f0[16] JB_ALIGNED, g0[16] JB_ALIGNED;
   float xm, dx;
   int i, j, n1;
-  n1 = n - 1;
+  n1 = (int) n - 1;
   xm = 0.5f * (xmin + xmax);
   dx = (xmax - xm) / (float) n1;
-  for (i = 0; i < n;)
+  for (i = 0; i < (int) n;)
     {
       for (j = 0; j < 16; ++j, ++i)
         {
@@ -2470,10 +2470,10 @@ check_uni_8xf64 (__m512d (*f) (const __m512d), double (*g) (const double),
   double x[8] JB_ALIGNED, f0[8] JB_ALIGNED, g0[8] JB_ALIGNED;
   double xm, dx;
   int i, j, n1;
-  n1 = n - 1;
+  n1 = (int) n - 1;
   xm = 0.5 * (xmin + xmax);
   dx = (xmax - xm) / (double) n1;
-  for (i = 0; i < n;)
+  for (i = 0; i < (int) n;)
     {
       for (j = 0; j < 8; ++j, ++i)
         {
@@ -2589,10 +2589,10 @@ check_uni_4xf32 (float32x4_t (*f) (const float32x4_t), float (*g) (const float),
   float x[4] JB_ALIGNED, f0[4] JB_ALIGNED, g0[4] JB_ALIGNED;
   float xm, dx;
   int i, j, n1;
-  n1 = n - 1;
+  n1 = (int) n - 1;
   xm = 0.5f * (xmin + xmax);
   dx = (xmax - xm) / (float) n1;
-  for (i = 0; i < n;)
+  for (i = 0; i < (int) n;)
     {
       for (j = 0; j < 4; ++j, ++i)
         {
@@ -2704,10 +2704,10 @@ check_uni_2xf64 (float64x2_t (*f) (const float64x2_t),
   double x[2] JB_ALIGNED, f0[2] JB_ALIGNED, g0[2] JB_ALIGNED;
   double xm, dx;
   int i, j, n1;
-  n1 = n - 1;
+  n1 = (int) n - 1;
   xm = 0.5 * (xmin + xmax);
   dx = (xmax - xm) / (double) n1;
-  for (i = 0; i < n;)
+  for (i = 0; i < (int) n;)
     {
       for (j = 0; j < 2; ++j, ++i)
         {
@@ -2826,10 +2826,10 @@ check_uni_4xf32 (vfloat32m1_t (*f) (const vfloat32m1_t),
   float x[4] JB_ALIGNED, f0[4] JB_ALIGNED, g0[4] JB_ALIGNED;
   float xm, dx;
   int i, j, n1;
-  n1 = n - 1;
+  n1 = (int) n - 1;
   xm = 0.5f * (xmin + xmax);
   dx = (xmax - xm) / (float) n1;
-  for (i = 0; i < n;)
+  for (i = 0; i < (int) n;)
     {
       for (j = 0; j < 4; ++j, ++i)
         {
@@ -2942,10 +2942,10 @@ check_uni_2xf64 (vfloat64m1_t (*f) (const vfloat64m1_t),
   double x[2] JB_ALIGNED, f0[2] JB_ALIGNED, g0[2] JB_ALIGNED;
   double xm, dx;
   int i, j, n1;
-  n1 = n - 1;
+  n1 = (int) n - 1;
   xm = 0.5 * (xmin + xmax);
   dx = (xmax - xm) / (double) n1;
-  for (i = 0; i < n;)
+  for (i = 0; i < (int) n;)
     {
       for (j = 0; j < 2; ++j, ++i)
         {
@@ -3060,10 +3060,10 @@ check_uni_8xf32 (vfloat32m1_t (*f) (const vfloat32m1_t),
   float x[8] JB_ALIGNED, f0[8] JB_ALIGNED, g0[8] JB_ALIGNED;
   float xm, dx;
   int i, j, n1;
-  n1 = n - 1;
+  n1 = (int) n - 1;
   xm = 0.5f * (xmin + xmax);
   dx = (xmax - xm) / (float) n1;
-  for (i = 0; i < n;)
+  for (i = 0; i < (int) n;)
     {
       for (j = 0; j < 8; ++j, ++i)
         {
@@ -3176,10 +3176,10 @@ check_uni_4xf64 (vfloat64m1_t (*f) (const vfloat64m1_t),
   double x[4] JB_ALIGNED, f0[4] JB_ALIGNED, g0[4] JB_ALIGNED;
   double xm, dx;
   int i, j, n1;
-  n1 = n - 1;
+  n1 = (int) n - 1;
   xm = 0.5 * (xmin + xmax);
   dx = (xmax - xm) / (double) n1;
-  for (i = 0; i < n;)
+  for (i = 0; i < (int) n;)
     {
       for (j = 0; j < 4; ++j, ++i)
         {
@@ -3295,10 +3295,10 @@ check_uni (JBFLOAT (*f) (const JBFLOAT), JBFLOAT (*g) (const JBFLOAT),
 {
   JBFLOAT x, xm, dx, f0, g0;
   int i, n1;
-  n1 = n - 1;
+  n1 = (int) n - 1;
   xm = (JBFLOAT) 0.5 *(xmin + xmax);
   dx = (xmax - xm) / (JBFLOAT) n1;
-  for (i = 0; i < n; ++i)
+  for (i = 0; i < (int) n; ++i)
     {
       if (i == 0)
         x = xmin;
@@ -3371,10 +3371,10 @@ check_unil (JBDOUBLE (*f) (const JBDOUBLE), JBDOUBLE (*g) (const JBDOUBLE),
 {
   JBDOUBLE x, xm, dx, f0, g0;
   int i, n1;
-  n1 = n - 1;
+  n1 = (int) n - 1;
   xm = (JBDOUBLE) 0.5 *(xmin + xmax);
   dx = (xmax - xm) / (JBDOUBLE) n1;
-  for (i = 0; i < n; ++i)
+  for (i = 0; i < (int) n; ++i)
     {
       if (i == 0)
         x = xmin;
@@ -6334,7 +6334,7 @@ main (void)
 
   printf ("check farray functions\n");
   fa = jbm_farray_create (pf, 11);
-  printf ("fa=%llu fa->x=%llu\n", (size_t) fa, (size_t) fa->x);
+  printf ("fa=%zu fa->x=%zu\n", (size_t) fa, (size_t) fa->x);
   fa2 = jbm_farray_new (11);
   fa3 = jbm_farray_new (11);
   for (i = 0; i < 11; ++i)
